Add files via upload
This project discusses alternating electric current three phases transmission using optical fibers for three optical channels separated with 450nm 540nm 630nm. The LED incorporates alternating current into it by using Direct Modulation, Then the alternating electric current is sent in the form of a light pulse. Then, the three channels are sent through a plastic optical fiber with 10m for all channels. The power was measured for all three wavelengths before and after the electric current was modulated. The results were taken and analyzed in a practical way, and the system was designed using a program EasyEDA, different parameters have been investigated which are the voltage difference of the alternating current that was included for the three LEDs and the power for all LEDs before connecting them to the optical fiber and after connecting them to the optical fiber.